+++ src/main.rs
@@ -74,7 +74,7 @@ mod test {
 
     #[test]
     fn try_insight() {
-        use voyager_client::insight::*;
+        use voyager_client::insight_client::*;
 
         let base = InsightWeather::new();
         base.query().unwrap();
@@ -98,5 +98,28 @@ mod test {
 
         let key = from_dotenv().unwrap();
         println!("{}", key);
+    }
+
+    #[test]
+    fn try_tech_transfer_patent() {
+        use voyager_client::tech_transfer::*;
+
+        let base = TechTransferClient::new();
+
+        let query = String::from("engine");
+        let res = base.query(query).unwrap();
+        println!("{}", res);
+    }
+
+    #[test]
+    fn try_tech_transfer_software() {
+        use voyager_client::tech_transfer::*;
+
+        let mut base = TechTransferClient::new();
+        base.software();
+
+        let query = String::from("engine");
+        let res = base.query(query).unwrap();
+        println!("{}", res);
     }
 }